backend/requirements.txt
torch>=2.0.0
torchvision
transformers>=4.30.0
datasets>=2.0.0
peft>=0.18.0
accelerate
sentencepiece
Pillow
flask
flask-cors
tqdm
scikit-learn
numpy
pandas
opencv-python
requests
